+++ AddOns/NugRunning/CHANGELOG.md
@@ -1,6 +1,8 @@
 # NugRunning
 
-## [9.0.2](https://github.com/rgd87/NugRunning/tree/9.0.2) (2020-10-15)
-[Full Changelog](https://github.com/rgd87/NugRunning/compare/9.0.1...9.0.2) [Previous Releases](https://github.com/rgd87/NugRunning/releases)
+## [9.0.5](https://github.com/rgd87/NugRunning/tree/9.0.5) (2020-11-29)
+[Full Changelog](https://github.com/rgd87/NugRunning/compare/9.0.4...9.0.5) [Previous Releases](https://github.com/rgd87/NugRunning/releases)
 
-- Fixed UA timer not working when Rampant Afflictions is picked  
+- Updated TOC  
+- Fixed bug with talent adjustments and disabled overlay  
+- Rampage timer can be disabled at initialization